Move Python CLI to cli.py
